Finish processing using ClauseIE
Need to touch up the final output forms, mostly by checking if there are
phrasal form need to be included (phrasal proposition for example). Need
to do analysis on about 1000 sentences to ensure output quality.

Need to create the simplest baseline to do real testing. The simplest one
would be to just generate sentences independently (sentences' order in a
movie is not taken into account)